+++ changelog.md
@@ -1,5 +1,26 @@
 # Changelog
 
+### v0.9.3, 2017-11-19
+ * fixed: don't switch active tab when mute/unmute tabs via a click on the sound button
+
+
+***
+
+
+### v0.9.2, 2017-11-16
+ * fixed: when the sidebar gets first opened no tab is selected (#148)
+
+
+***
+
+
+### v0.9.1, 2017-11-14
+ * fixed: typo in the update landing page
+
+
+***
+
+
 ### v0.9.0, 2017-11-13
  * this is the first full WebExtension version (#13) 🤖
  * added: landing page after update with changes explained (#13)